+++ Config/sql_files/aggregate/resources_counts_grouped_by_account_id.sql
@@ -1,0 +1,5 @@
+SELECT
+  COUNT(*),
+  accountId
+GROUP BY
+  accountId